requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
matplotlib>=3.4.3
scikit-learn>=0.24.2
tqdm>=4.62.3
Pillow>=8.3.2
kaggle>=1.5.0 